[SUMMARY]
Added in a new method, DoUntilSucceeds() that just tries to do a Do() until the redis server accepts it. Will only retry if connection pool is exhausted, not if the query is invalid.

[ISSUES]

None

[NOTES]

Doesn't have a mechanism to stop pushing to redis: will run forever if redis is down or if there's another issue that might cause it to fail.